HSEARCH-3797 Move collector implementations to the "lowlevel" package
So that they can be used from other packages without introducing
circular dependencies between packages.

We will need this in the following commits, though we will only
use a few of these classes.
