+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/lowlevel/collector/impl/LuceneCollectorFactory.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.search.extraction.impl;
+package org.hibernate.search.backend.lucene.lowlevel.collector.impl;
 
 import org.apache.lucene.facet.FacetsCollector;
 import org.apache.lucene.search.Collector;
