+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/lowlevel/collector/impl/LuceneCollectorKey.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.search.extraction.impl;
+package org.hibernate.search.backend.lucene.lowlevel.collector.impl;
 
 import org.apache.lucene.search.Collector;
 import org.apache.lucene.search.TopDocsCollector;
@@ -12,9 +12,10 @@ import org.apache.lucene.search.TotalHitCountCollector;
 
 /**
  * Tagging interface for collector keys.
+ * <p>
+ * This is used for de-duplication of collectors, to avoid collecting the same data twice during the same search.
+ *
  * @param <C> The type of collector.
- * @see ExtractionRequirements.Builder#requireCollector(LuceneCollectorFactory)
- * @see LuceneCollectors#getCollectors()
  */
 public interface LuceneCollectorKey<C extends Collector> {
