+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/projection/impl/LuceneEntityProjection.java
@@ -8,7 +8,7 @@ package org.hibernate.search.backend.lucene.search.projection.impl;
 
 import java.util.Set;
 
-import org.hibernate.search.backend.lucene.search.extraction.impl.DocumentReferenceCollector;
+import org.hibernate.search.backend.lucene.lowlevel.collector.impl.DocumentReferenceCollector;
 import org.hibernate.search.backend.lucene.search.extraction.impl.LuceneResult;
 import org.hibernate.search.engine.backend.common.DocumentReference;
 import org.hibernate.search.engine.search.loading.spi.LoadingResult;
